Init HitboxComponent, rename current Attack to Attack_old
Attack is being renamed to old since I want to redo part of the system. Instead of Attack having all of the info needed for an attack at all time, it will only hold the info needed at the time of impact. This allows Projectiles, Melee, and Aura to have their own dedicated classes