Add enemy spawn - press enter to spawn
NOTE: setup for wave spawning and logic is implemented, but not complete yet

Introduces SpawnerPath node and script to handle enemy wave spawning along a path, with logic to avoid spawning near the player. Updates input actions for movement and jumping, and removes static mob instances from the main game scene in favor of dynamic spawning.